separate CSV creation and word data extraction
